add option side type, add bitcoin spot price getter
